cpu_emulate.c: Add support for SUB $8,ESP.
cpu_emulate.c: Fix problem with SUB $12,ESP instruction.

systrace.c: Fix for sys_clone - now we don't get the core dumps.

systrace.c: Fix for sigreturn needing a proper call frame so it can patch the registers [i386 only]. This appears to work:

        dtrace -n syscall:::

cpu_emulate.c: Emulate PUSH ESI. We need to avoid having a zillion case/defines where we can use the CPU opcode decoding to get big-bang-for-buck. We need to pass in the missing opcode so we can decide which is better to do.